feat: CPU 타입을 double로 변경

+++ back-end/src/main/java/site/yacht/backend/domain/deployment/api/response/DeployDetailResponse.java
@@ -15,7 +15,7 @@ public class DeployDetailResponse {
     private String gitUrl;
     private String namespace;
     private int replicaNumber;
-    private int cpu;
+    private double cpu;
     private int memory;
     private int port;
